requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
typing-extensions
python-multipart